Add configuration attribute Crono::CronoJob.maintain_log for conditionally excluding logging
When running jobs every minute, for example, logging to the database can
become problematic as it will baloon in size and result in some
extremely large UPDATE statements.

This configuration option allows users to conditionally disable this by
setting:

    Crono::CronoJob.maintain_log = false

With this in place, the crono_jobs entries still get updated, but
exclude the log attribute.